# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowTensor API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ParseRequest {
  /** Python source code to parse (PyTorch or Pandas) */
  code: string;
}

/**
 * Category of the node
 */
export type FlowNodeType = (typeof FlowNodeType)[keyof typeof FlowNodeType];

export const FlowNodeType = {
  pandas: "pandas",
  pytorch: "pytorch",
  input: "input",
  output: "output",
  intermediate: "intermediate",
} as const;

export interface FlowNode {
  /** Unique node identifier */
  id: string;
  /** Category of the node */
  type: FlowNodeType;
  /** Short display name for the node (e.g., "dropna()") */
  label: string;
  /** Plain-English explanation of what the operation does */
  description: string;
  /** Inferred input shape/size (e.g., "1000x5", "[32, 128]") */
  input_shape: string | null;
  /** Inferred output shape/size (e.g., "800x3", "[32, 64]") */
  output_shape: string | null;
  /** Suggested x position for the node in the graph */
  position_x: number;
  /** Suggested y position for the node in the graph */
  position_y: number;
  /** Optional group container name (e.g., class body, loop body) — frontend can render these as React Flow subflows */
  group?: string | null;
}

export interface FlowEdge {
  /** Unique edge identifier */
  id: string;
  /** Source node id */
  source: string;
  /** Target node id */
  target: string;
}

/**
 * Detected framework in the code
 */
export type ParseResponseFramework =
  (typeof ParseResponseFramework)[keyof typeof ParseResponseFramework];

export const ParseResponseFramework = {
  pandas: "pandas",
  pytorch: "pytorch",
  mixed: "mixed",
  unknown: "unknown",
} as const;

export interface ParseResponse {
  nodes: FlowNode[];
  edges: FlowEdge[];
  /** Detected framework in the code */
  framework: ParseResponseFramework;
}

export interface FetchGistRequest {
  /** A public GitHub Gist URL */
  url: string;
}

export interface GistFile {
  filename: string;
  /** File size in bytes */
  size: number;
}

export interface GistResult {
  /** Raw file content. Present when a single .py file is auto-selected. */
  code: string | null;
  /** Filename of the auto-selected file. Present alongside `code`. */
  filename: string | null;
  /** List of .py files when multiple are present and the user must choose. */
  files: GistFile[] | null;
}

export interface ParseError {
  /** Human-friendly error message */
  error: string;
  /** Technical detail about the parse failure */
  detail: string | null;
}