# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowTensor API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: parse
    description: Python code parsing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /parse:
    post:
      operationId: parseCode
      tags: [parse]
      summary: Parse Python code into a flow graph
      description: Parses PyTorch or Pandas code using AST analysis and returns nodes and edges representing the data flow graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseRequest"
      responses:
        "200":
          description: Successfully parsed code into a flow graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseResponse"
        "422":
          description: Code could not be parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseError"
  /fetch-gist:
    post:
      operationId: fetchGist
      tags: [parse]
      summary: Fetch a public GitHub Gist and return its Python files
      description: Accepts a Gist URL, fetches the public Gist via the GitHub API, and returns its Python file(s).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchGistRequest"
      responses:
        "200":
          description: Gist fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GistResult"
        "422":
          description: Could not fetch the Gist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ParseRequest:
      type: object
      properties:
        code:
          type: string
          description: Python source code to parse (PyTorch or Pandas)
      required:
        - code
    FlowNode:
      type: object
      properties:
        id:
          type: string
          description: Unique node identifier
        type:
          type: string
          enum: [pandas, pytorch, input, output, intermediate]
          description: Category of the node
        label:
          type: string
          description: Short display name for the node (e.g., "dropna()")
        description:
          type: string
          description: Plain-English explanation of what the operation does
        input_shape:
          type: string
          description: Inferred input shape/size (e.g., "1000x5", "[32, 128]")
          nullable: true
        output_shape:
          type: string
          description: Inferred output shape/size (e.g., "800x3", "[32, 64]")
          nullable: true
        position_x:
          type: number
          description: Suggested x position for the node in the graph
        position_y:
          type: number
          description: Suggested y position for the node in the graph
        group:
          type: string
          description: Optional group container name (e.g., class body, loop body) — frontend can render these as React Flow subflows
          nullable: true
      required:
        - id
        - type
        - label
        - description
        - input_shape
        - output_shape
        - position_x
        - position_y
    FlowEdge:
      type: object
      properties:
        id:
          type: string
          description: Unique edge identifier
        source:
          type: string
          description: Source node id
        target:
          type: string
          description: Target node id
      required:
        - id
        - source
        - target
    ParseResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/FlowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/FlowEdge"
        framework:
          type: string
          enum: [pandas, pytorch, mixed, unknown]
          description: Detected framework in the code
      required:
        - nodes
        - edges
        - framework
    FetchGistRequest:
      type: object
      properties:
        url:
          type: string
          description: A public GitHub Gist URL
      required:
        - url
    GistFile:
      type: object
      properties:
        filename:
          type: string
        size:
          type: integer
          description: File size in bytes
      required:
        - filename
        - size
    GistResult:
      type: object
      properties:
        code:
          type: string
          description: Raw file content. Present when a single .py file is auto-selected.
          nullable: true
        filename:
          type: string
          description: Filename of the auto-selected file. Present alongside `code`.
          nullable: true
        files:
          type: array
          description: List of .py files when multiple are present and the user must choose.
          nullable: true
          items:
            $ref: "#/components/schemas/GistFile"
      required:
        - code
        - filename
        - files
    ParseError:
      type: object
      properties:
        error:
          type: string
          description: Human-friendly error message
        detail:
          type: string
          description: Technical detail about the parse failure
          nullable: true
      required:
        - error
        - detail