# lib/api-zod/src/generated/types/fetchGistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowTensor API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FetchGistRequest {
  /** A public GitHub Gist URL */
  url: string;
}