# lib/api-zod/src/generated/types/gistFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowTensor API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GistFile {
  filename: string;
  /** File size in bytes */
  size: number;
}