# lib/api-zod/src/generated/types/gistResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlowTensor API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GistFile } from "./gistFile";

export interface GistResult {
  /** Raw file content. Present when a single .py file is auto-selected. */
  code: string | null;
  /** Filename of the auto-selected file. Present alongside `code`. */
  filename: string | null;
  /** List of .py files when multiple are present and the user must choose. */
  files: GistFile[] | null;
}